+++ perftimer.h
@@ -2,6 +2,6 @@
 
 void TimerInit(void);
 void Timer(void);
-unsigned long GetTimer(void);
-unsigned long GetTimerR(void);
+uint32_t GetTimer(void);
+uint32_t GetTimerR(void);
 void TimerTest();
